AsyncLocal in activity scope needs proper initialization (#298)
