+++ Sdk/Speckle.Connectors.Logging/ActivityScopeExtensions.cs
@@ -1,11 +1,8 @@
-﻿using System.Diagnostics;
-using OpenTelemetry;
-
-namespace Speckle.Connectors.Logging;
+﻿namespace Speckle.Connectors.Logging;
 
 public static class ActivityScope
 {
-  private static readonly AsyncLocal<Dictionary<string, object>> s_tags = new();
+  private static readonly AsyncLocal<Dictionary<string, object>> s_tags = new() { Value = new() };
 
   public static IReadOnlyDictionary<string, object> Tags => s_tags.Value;
 
@@ -18,16 +15,5 @@ public static class ActivityScope
   private sealed class TagScope(string key) : IDisposable
   {
     public void Dispose() => s_tags.Value.Remove(key);
-  }
-}
-
-internal sealed class ActivityScopeProcessor : BaseProcessor<Activity>
-{
-  public override void OnEnd(Activity data)
-  {
-    foreach (KeyValuePair<string, object> keyValuePair in ActivityScope.Tags)
-    {
-      data.SetTag(keyValuePair.Key, keyValuePair.Value);
-    }
   }
 }
